+++ chalice/pipeline.py
@@ -6,19 +6,20 @@ from chalice.config import Config  # noqa
 from chalice import constants
 
 
-def create_pipeline_template(config):
-    # type: (Config) -> Dict[str, Any]
-    pipeline = CreatePipelineTemplate()
-    return pipeline.create_template(config.app_name,
-                                    config.lambda_python_version)
-
-
 class InvalidCodeBuildPythonVersion(Exception):
     def __init__(self, version):
         # type: (str) -> None
         super(InvalidCodeBuildPythonVersion, self).__init__(
             'CodeBuild does not yet support python version %s.' % version
         )
+
+
+class PipelineParameters(object):
+    def __init__(self, app_name, lambda_python_version, codebuild_image=None):
+        # type: (str, str, Optional[str]) -> None
+        self.app_name = app_name
+        self.lambda_python_version = lambda_python_version
+        self.codebuild_image = codebuild_image
 
 
 class CreatePipelineTemplate(object):
@@ -45,29 +46,28 @@ class CreatePipelineTemplate(object):
         "Outputs": {},
     }
 
-    def __init__(self):
-        # type: () -> None
-        pass
-
-    def _codebuild_image(self, lambda_python_version):
-        # type: (str) -> str
-        try:
-            image_suffix = self._CODEBUILD_IMAGE[lambda_python_version]
-            return 'aws/codebuild/%s' % image_suffix
-        except KeyError as e:
-            raise InvalidCodeBuildPythonVersion(str(e))
-
-    def create_template(self, app_name, python_lambda_version):
-        # type: (str, str) -> Dict[str, Any]
+    def create_template(self, pipeline_params):
+        # type: (PipelineParameters) -> Dict[str, Any]
         t = copy.deepcopy(self._BASE_TEMPLATE)  # type: Dict[str, Any]
-        t['Parameters']['ApplicationName']['Default'] = app_name
-        t['Parameters']['CodeBuildImage']['Default'] = self._codebuild_image(
-            python_lambda_version)
+        params = t['Parameters']
+        params['ApplicationName']['Default'] = pipeline_params.app_name
+        params['CodeBuildImage']['Default'] = self._get_codebuild_image(
+            pipeline_params)
 
         resources = [SourceRepository, CodeBuild, CodePipeline]
         for resource_cls in resources:
             resource_cls().add_to_template(t)
         return t
+
+    def _get_codebuild_image(self, params):
+        # type: (PipelineParameters) -> str
+        if params.codebuild_image is not None:
+            return params.codebuild_image
+        try:
+            image_suffix = self._CODEBUILD_IMAGE[params.lambda_python_version]
+            return 'aws/codebuild/%s' % image_suffix
+        except KeyError as e:
+            raise InvalidCodeBuildPythonVersion(str(e))
 
 
 class BaseResource(object):