+++ tests/functional/cli/test_cli.py
@@ -311,6 +311,23 @@ def test_can_generate_pipeline_for_all(runner):
             assert "Outputs" in template
 
 
+def test_no_errors_if_override_codebuild_image(runner):
+    with runner.isolated_filesystem():
+        cli.create_new_project_skeleton('testproject')
+        os.chdir('testproject')
+        result = _run_cli_command(
+            runner, cli.generate_pipeline,
+            ['-i', 'python:3.6.1', 'pipeline.json'])
+        assert result.exit_code == 0, result.output
+        assert os.path.isfile('pipeline.json')
+        with open('pipeline.json', 'r') as f:
+            template = json.load(f)
+            # The actual contents are tested in the unit
+            # tests.  Just a sanity check that it looks right.
+            image = template['Parameters']['CodeBuildImage']['Default']
+            assert image == 'python:3.6.1'
+
+
 def test_env_vars_set_in_local(runner, mock_cli_factory,
                                monkeypatch):
     local_server = mock.Mock(spec=local.LocalDevServer)